refactor: move pages to root (#488)
