+++ packages/web/pages/500.tsx
