+++ packages/web/pages/earn/pool/remove.tsx
