+++ packages/web/pages/earn/pool/stake.tsx
